requirements.txt
Django>=4.2.0,<5.0
psycopg2-binary>=2.9.6
redis>=4.5.4
gunicorn>=20.1.0
Pillow>=9.5.0
django-redis>=5.2.0
django-allauth>=0.54.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-taggings>=1.5.0
django-ckeditor>=6.5.1
django-rest-framework>=3.14.0
drf-yasg>=1.21.5
python-dotenv>=1.0.0
factory-boy>=3.2.1
pytest-django>=4.5.2
django-debug-toolbar>=4.1.0
django-cleanup>=7.0.0
readtime
drf-spectacular
whitenoise[brotli]